Created subdirectories to further organize source files.
