+++ src/internal/allocators/alloc_stack.c
