+++ src/internal/procedures/differentiate.c
