+++ src/objects/packed_function.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "function.h"
-#include "../internal/alloc_stack.h"
+#include "../internal/allocators/alloc_stack.h"
 
 pf_object make_packed_copy(f_object *object);
