+++ src/readers/function_reader.h
@@ -1,8 +1,8 @@
 #pragma once
 
 #include "../objects/function.h"
-#include "../internal/alloc_stack.h"
 #include "../internal/token_list.h"
+#include "../internal/allocators/alloc_stack.h"
 #include "../objects/packed_function.h"
 #include "../env/commands.h"
